Migrate RuleHelper to typescript and functional component
